gajim-otr: fixed broken xbps-src fetch
The gajim web server seems to regenerate its plugin's zip files
frequently (hourly?), and the file order and/or other properties seem
to be indeterminate.

The bottom line is that the contents constantly change, so the
xbps-src digest check fails, even though the actual contents of the
archive haven't changed.

This adds a custom fetch function, which checksums the files in a
reproducible fashion. It's rather a kludge, but seems to work.